Class proxy setup
The data structures for items have been overhauled as a new method of managing data has been introduced.

Based off the system in place for cyberpunk red, each item / actor can now be wrapped in its own custom class through the use of a proxy trap on the constructor. This means all the item type management and sorting that happens internally can be removed, and the individual functions and set up can be moved to their respective unique classes